#155 Early work for new BitSet-based column management

+++ src/tilda/types/Type_BitfieldCollectionNull.java
@@ -18,7 +18,7 @@ package tilda.types;
 
 public class Type_BitfieldCollectionNull extends Type_BitfieldCollection implements Nullable   
   {
-    public Type_BitfieldCollectionNull(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_BitfieldCollectionNull(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, Description);
       }
