+++ src/tilda/types/Type_BitfieldPrimitive.java
@@ -20,7 +20,7 @@ import tilda.enums.ColumnType;
 
 public class Type_BitfieldPrimitive extends ColumnDefinition
   {
-    public Type_BitfieldPrimitive(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_BitfieldPrimitive(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, ColumnType.BITFIELD, false, Description);
       }
